--- src/MyHomeworks/Homework1/MyClass.java
@@ -1,13 +1,0 @@
-package MyHomeworks.Homework1;
-
-class MyClass {
-
-    public static int sumOfTwoNumbers(int a, int b) {
-        return a + b;
-    }
-
-    public static void main(String[] args) {
-        int c = sumOfTwoNumbers(100, 200);
-        System.out.println(c);
-    }
-}
