--- src/MyHomeworks/Homework5/Task2/Test.java
@@ -1,13 +1,0 @@
-package MyHomeworks.Homework5.Task2;
-
-public class Test {
-    public static void main(String[] args) {
-        Dog dog = new Dog();
-        int paws = dog.paw;
-        System.out.println("Number of paws: " + paws);
-
-        Cat cat = new Cat();
-        cat.sleep();
-
-    }
-}
