+++ src/main/java/MyHomeworks/Main.java
@@ -1,0 +1,7 @@
+package MyHomeworks;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
